+++ argilla-frontend/components/features/global/user/UserAvatarTooltip.vue
@@ -1,6 +1,11 @@
 <template>
   <div v-if="$auth.loggedIn" v-click-outside="close" class="user">
-    <a class="user__button" @click.prevent="showSelector">
+    <a
+      class="user__button"
+      @click.prevent="showSelector"
+      role="button"
+      aria-label="User Button"
+    >
       {{ user.avatar }}
     </a>
     <div v-if="visibleSelector && user" class="user__content">
@@ -17,7 +22,7 @@
           </div>
         </div>
       </div>
-      <div class="">
+      <div class="" role="list">
         <NuxtLink class="user__link" :to="{ name: 'user-settings' }">
           {{ $t("userAvatarTooltip.settings") }}
         </NuxtLink>