+++ argilla-frontend/translation/en.js
@@ -257,6 +257,77 @@ export default {
       team: "Team progress",
     },
   },
+  home: {
+    argillaDatasets: "Your datasets",
+    none: "None yet",
+    importTitle: "Import a dataset from Hugging Face Hub",
+    importText:
+      "Start with a dataset from the Hub by simply pasting the repository name",
+    importButton: "Import dataset",
+    importFromHub: "Import dataset from Hugging Face",
+    importFromPython: "Import from Python",
+    importFromPythonHFWarning:
+      "If you're using a private Space, check the <a target='_blank' href='https://docs.argilla.io/latest/getting_started/how-to-configure-argilla-on-huggingface/#how-to-use-private-spaces'>docs</a>.",
+    exampleDatasetsTitle: "Don’t know where to start?",
+    exampleDatasetsText: "Explore these example datasets",
+    guidesTitle: "Not familiar with Argilla?",
+    guidesText: "Take a look at these guides:",
+    pasteRepoIdPlaceholder: "Paste a repo id e.g., stanfordnlp/imdb",
+    demoLink:
+      "Log into this <a href='https://huggingface.co/spaces/argilla/argilla-template-space' target='_blank'>demo</a> to try Argilla out",
+  },
+  datasetCreation: {
+    questions: {
+      labelSelection: {
+        atLeastTwoOptions: "At least two options are required",
+        optionsWithoutLabel: "Empty options are not allowed",
+      },
+    },
+    atLeastOneQuestion: "At least one question is required.",
+    atLeastOneRequired: "At least one required question is needed.",
+    hasInvalidQuestions: "Some questions are invalid",
+    createDataset: "Create the dataset in Argilla",
+    datasetName: "Dataset name",
+    name: "Name",
+    assignWorkspace: "Assign a workspace",
+    selectSplit: "Select a split",
+    recordWarning:
+      "The created dataset will include the first 10K rows and further records can be logged via the python SDK.",
+    button: "Create dataset",
+    fields: "Fields",
+    questionsTitle: "Questions",
+    yourQuestions: "Your questions",
+    requiredField: "Required field",
+    requiredQuestion: "Required question",
+    select: "Select",
+    mapToColumn: "Map to column",
+    subset: "Subset",
+    selectSubset: "Your can create a dataset from only one subset.",
+    preview: "Preview",
+    importData: "Import data",
+    addRecords: "Add records",
+    cantLoadRepository: "Dataset not found or available on Hugging Face",
+    none: "None",
+    noWorkspaces:
+      "Please, follow this <a target='_blank' href='https://docs.argilla.io/latest/how_to_guides/workspace/#create-a-new-workspace'>guide</a> to create a workspace",
+  },
+  config: {
+    field: {
+      text: "Text field",
+      chat: "Chat field",
+      image: "Image field",
+      "no mapping": "No mapping",
+    },
+    question: {
+      text: "Text",
+      rating: "Rating",
+      label_selection: "Label",
+      ranking: "Ranking",
+      multi_label_selection: "Multi-label",
+      span: "Span",
+      "no mapping": "No mapping",
+    },
+  },
   persistentStorage: {
     adminOrOwner:
       "Persistent storage is not enabled. All data will be lost if this space restarts. Go to the space settings to enable it.",