+++ argilla-server/src/argilla_server/api/handlers/v1/datasets/datasets.py
@@ -29,6 +29,7 @@ from argilla_server.api.schemas.v1.datasets import (
     DatasetProgress,
     Datasets,
     DatasetUpdate,
+    HubDataset,
     UsersProgress,
 )
 from argilla_server.api.schemas.v1.fields import Field, FieldCreate, Fields
@@ -38,9 +39,11 @@ from argilla_server.api.schemas.v1.metadata_properties import (
     MetadataPropertyCreate,
 )
 from argilla_server.api.schemas.v1.vector_settings import VectorSettings, VectorSettingsCreate, VectorsSettings
+from argilla_server.api.schemas.v1.jobs import Job as JobSchema
 from argilla_server.contexts import datasets
 from argilla_server.database import get_async_db
 from argilla_server.enums import DatasetStatus
+from argilla_server.jobs import hub_jobs
 from argilla_server.models import Dataset, User
 from argilla_server.search_engine import (
     SearchEngine,
@@ -152,7 +155,7 @@ async def get_current_user_dataset_metrics(
 
     await authorize(current_user, DatasetPolicy.get(dataset))
 
-    result = await datasets.get_user_dataset_metrics(db, search_engine, current_user, dataset)
+    result = await datasets.get_user_dataset_metrics(search_engine, current_user, dataset)
 
     return DatasetMetrics(responses=result)
 
@@ -308,3 +311,26 @@ async def update_dataset(
     await authorize(current_user, DatasetPolicy.update(dataset))
 
     return await datasets.update_dataset(db, dataset, dataset_update.dict(exclude_unset=True))
+
+
+@router.post("/datasets/{dataset_id}/import", status_code=status.HTTP_202_ACCEPTED, response_model=JobSchema)
+async def import_dataset_from_hub(
+    *,
+    db: AsyncSession = Depends(get_async_db),
+    dataset_id: UUID,
+    hub_dataset: HubDataset,
+    current_user: User = Security(auth.get_current_user),
+):
+    dataset = await Dataset.get_or_raise(db, dataset_id)
+
+    await authorize(current_user, DatasetPolicy.import_from_hub(dataset))
+
+    job = hub_jobs.import_dataset_from_hub_job.delay(
+        name=hub_dataset.name,
+        subset=hub_dataset.subset,
+        split=hub_dataset.split,
+        dataset_id=dataset.id,
+        mapping=hub_dataset.mapping.dict(),
+    )
+
+    return JobSchema(id=job.id, status=job.get_status())